clear

+++ data-mappers/dto-to-domain/src/commonMain/kotlin/com/grippo/dto/domain/achievement/AchievementMapper.kt
@@ -60,7 +60,7 @@ private fun BestWeight.toDomainOrNull(): Achievement.BestWeight? {
         exerciseExampleId = exampleId,
         exerciseId = exerciseId,
         iterationId = iterationId,
-        weight = weight
+        weight = weight.toInt()
     )
 }
